[Behavioral Analytics] refactor withLensData hoc
