+++ x-pack/plugins/enterprise_search/public/applications/analytics/components/analytics_overview/analytics_collection_table.tsx
@@ -25,11 +25,11 @@ import { OnTimeChangeProps } from '@elastic/eui/src/components/date_picker/super
 import { i18n } from '@kbn/i18n';
 
 import { AnalyticsCollection } from '../../../../../common/types/analytics';
+import { FilterBy } from '../../utils/get_formula_by_filter';
 import { AddAnalyticsCollection } from '../add_analytics_collections/add_analytics_collection';
 
 import { AnalyticsCollectionCardWithLens } from './analytics_collection_card/analytics_collection_card';
 
-import { FilterBy } from './analytics_collection_card/with_lens_data';
 import { AnalyticsCollectionTableStyles } from './analytics_collection_table.styles';
 
 const defaultQuickRanges: EuiSuperDatePickerCommonRange[] = [
@@ -98,7 +98,7 @@ export const AnalyticsCollectionTable: React.FC<AnalyticsCollectionTableProps> =
     ],
     [analyticsCollectionTableStyles.button]
   );
-  const [filterId, setFilterId] = useState<string>(filterOptions[0].id);
+  const [filterId, setFilterId] = useState<FilterBy>(filterOptions[0].id);
   const [timeRange, setTimeRange] = useState<{ from: string; to: string }>({
     from: defaultQuickRanges[0].start,
     to: defaultQuickRanges[0].end,
@@ -127,7 +127,7 @@ export const AnalyticsCollectionTable: React.FC<AnalyticsCollectionTableProps> =
           <EuiFlexItem grow={false}>
             <EuiButtonGroup
               css={analyticsCollectionTableStyles.buttonGroup}
-              onChange={setFilterId}
+              onChange={(newFilterId) => setFilterId(newFilterId as FilterBy)}
               color="primary"
               buttonSize="compressed"
               idSelected={filterId}
@@ -156,6 +156,7 @@ export const AnalyticsCollectionTable: React.FC<AnalyticsCollectionTableProps> =
         {collections.map((collection) => (
           <AnalyticsCollectionCardWithLens
             key={collection.name}
+            id={`collection-card-${collection.name}`}
             collection={collection}
             subtitle={selectedFilterLabel}
             filterBy={filterId}
