+++ x-pack/plugins/enterprise_search/public/applications/analytics/utils/get_formula_by_filter.ts
@@ -1,0 +1,23 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+export enum FilterBy {
+  Searches = 'Searches',
+  NoResults = 'NoResults',
+  Clicks = 'Clicks',
+  Sessions = 'Sessions',
+}
+export const getFormulaByFilter = (filter: FilterBy, shift?: string | null): string => {
+  const mapFilterByToFormula: { [key in FilterBy]: string } = {
+    [FilterBy.Searches]: "count(kql='event.action: search'",
+    [FilterBy.NoResults]: "count(kql='event.customer_data.totalResults : 0'",
+    [FilterBy.Clicks]: 'count(',
+    [FilterBy.Sessions]: 'count(',
+  };
+
+  return mapFilterByToFormula[filter] + (shift ? `, shift='${shift}'` : '') + ')';
+};
